Remove ethical metrics randomize
