+++ packages/dappmanager/src/utils/autoUpdateHelper.ts
@@ -12,7 +12,6 @@ import {
   AutoUpdatePending,
   AutoUpdateFeedback
 } from "@dappnode/common";
-import { getRandomizedInterval } from "./getRandomizedInterval.js";
 
 // Groups of packages keys
 export const MY_PACKAGES = "my-packages";
@@ -20,6 +19,24 @@ export const SYSTEM_PACKAGES = "system-packages";
 
 const updateDelay = getRandomizedInterval(params.AUTO_UPDATE_DELAY, params.AUTO_UPDATE_DAEMON_INTERVAL);
 const coreDnpName = params.coreDnpName;
+
+/**
+ * Randomize an interval
+ *
+ * Example:
+ * getRandomizedInterval(50, 10) // 50 +/- 10 = [40, 60]
+ *
+ * @param baseInterval
+ * @param variation
+ * @returns
+ */
+export function getRandomizedInterval(
+  baseInterval: number,
+  variation: number
+): number {
+  const randomAdjustment = Math.round((Math.random() * 2 - 1) * variation); // Random integer between -variation and +variation
+  return baseInterval + randomAdjustment;
+}
 
 /**
  * Define types
