+++ packages/params/src/params.ts
@@ -135,8 +135,7 @@ export const params = {
   CHECK_DISK_USAGE_DAEMON_INTERVAL: 1 * MINUTE,
   NAT_RENEWAL_DAEMON_INTERVAL: 1 * HOUR,
   NSUPDATE_DAEMON_INTERVAL: 1 * HOUR,
-  ETH_METRICS_DAEMON_INTERVAL: 50 * MINUTE,
-  ETH_METRICS_DAEMON_INTERVAL_VARIATION: 10 * MINUTE,
+  ETHICAL_METRICS_DAEMON_INTERVAL: 50 * MINUTE,
 
   // IPFS parameters
   IPFS_HOST: process.env.IPFS_HOST || process.env.IPFS_REDIRECT,
